Assert that client-id and client-secret are not nil
Improves error messages when the client-id or client-secret is missing
from a profile.

This check does not try to validate the whole profile map, but just
concentrates on `client-secret` and `client-id` because these two
values are commonly taken from environment variables, so they are much
more likely to be nil by mistake than others.

Fixes #19.